feat: add next and remove react-router
