+++ package.json
@@ -4,25 +4,27 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "lint": "eslint . --ext ts,tsx",
     "lint-fix": "eslint --fix",
     "fix": "prettier --write .",
     "preview": "vite preview",
     "prepare": "husky install",
-    "test": "vitest run --coverage",
+    "test": "vitest run --coverage.enabled --coverage.all",
     "vi": "vitest"
   },
   "dependencies": {
     "@reduxjs/toolkit": "1.9.7",
     "axios": "1.5.1",
     "classnames": "2.3.2",
+    "next": "^14.0.2",
+    "next-redux-wrapper": "8.1.0",
     "query-string": "8.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-redux": "8.1.3",
-    "react-router-dom": "6.16.0",
     "uuid": "9.0.1"
   },
   "devDependencies": {
