+++ src/UI/Pagination/Pagination.module.css
