--- src/app/components/Search/Search.module.css
@@ -1,64 +1,0 @@
-:root {
-  --color-green: rgb(7, 202, 0);
-  --color-red-dark: rgb(179, 21, 0);
-}
-
-.search {
-  padding-block: 100px;
-  background-color: blueviolet;
-}
-
-.searchHeading {
-  inline-size: fit-content;
-  margin-block-end: 30px;
-  margin-inline: auto;
-}
-
-.searchForm {
-  margin-block: 20px;
-  margin-inline: auto;
-  max-inline-size: 400px;
-  display: flex;
-  justify-content: center;
-}
-
-.searchInput {
-  inline-size: 300px;
-  font-size: 30px;
-  border-block: 1px solid orange;
-  border-inline: 1px solid orange;
-  border-radius: 5px 0 0 5px;
-}
-
-.searchButton {
-  padding-block: 10px;
-  padding-inline: 10px;
-  background-color: var(--color-green);
-  border-block: 1px solid red;
-  border-inline: 1px solid red;
-  border-inline-end: none;
-  border-radius: 0 5px 5px 0;
-  font-size: 15px;
-  cursor: pointer;
-
-  transition: all 0.5s;
-}
-
-.searchButton:hover {
-  color: white;
-  background-color: var(--color-red-dark);
-}
-
-.searchError {
-  padding-block: 10px;
-  padding-inline: 10px;
-  background-color: red;
-  border-block: none;
-  border-inline: none;
-  border-radius: 0 5px 5px 0;
-  font-size: 15px;
-  cursor: pointer;
-  text-wrap: nowrap;
-
-  transition: all 0.5s;
-}
