--- src/app/store/store.ts
@@ -1,23 +1,0 @@
-import { combineReducers, configureStore } from '@reduxjs/toolkit';
-import type { PreloadedState } from '@reduxjs/toolkit';
-import query from './reducers/querySlice';
-import search from './reducers/searchSlice';
-import { animeApi } from './reducers/animeApi';
-
-const rootReducer = combineReducers({
-  query,
-  search,
-  [animeApi.reducerPath]: animeApi.reducer,
-});
-
-export const setupStore = (preloadedState?: PreloadedState<RootState>) => {
-  return configureStore({
-    reducer: rootReducer,
-    middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(animeApi.middleware),
-    preloadedState,
-  });
-};
-
-export type RootState = ReturnType<typeof rootReducer>;
-export type AppStore = ReturnType<typeof setupStore>;
-export type AppDispatch = AppStore['dispatch'];
