+++ src/components/AnimeItem/AnimeItem.tsx
@@ -1,0 +1,23 @@
+import { Anime } from '@tutkli/jikan-ts';
+
+import styles from './AnimeItem.module.css';
+import { useParamsNavigator } from '../../hooks/useNavigator';
+
+interface IAnimeItemProps {
+  anime: Anime;
+}
+
+export function AnimeItem({ anime }: IAnimeItemProps) {
+  const navigate = useParamsNavigator();
+
+  const onClickHandler = (e: React.MouseEvent) => {
+    e.stopPropagation();
+    navigate('details', null, String(anime.mal_id));
+  };
+
+  return (
+    <div className={styles.anime} onClick={onClickHandler}>
+      <h3 className={styles.anime__heading}>{anime.title_english || anime.title}</h3>
+    </div>
+  );
+}
