+++ src/components/DetailedAnimeItem/DetailedAnimeItem.module.css
