+++ src/components/DetailedAnimeItem/DetailedAnimeItem.tsx
