+++ src/components/NotFound/NotFound.module.css
@@ -1,0 +1,34 @@
+/* :root {
+  --color-yellow: rgb(224, 224, 1);
+} */
+
+.notFound {
+  background-color: black;
+  inline-size: 100%;
+  min-block-size: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  color: rgb(224, 224, 1);
+}
+
+.notFound__button {
+  font-size: 20px;
+  display: block;
+  margin-block-start: 50px;
+  color: white;
+  background-color: red;
+  border-block: none;
+  border-inline: none;
+  padding-block: 10px;
+  padding-inline: 25px;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: 0.4s ease;
+}
+
+.notFound__button:hover {
+  background-color: greenyellow;
+  color: red;
+}
