+++ src/components/Search/searchOptions.ts
