+++ src/styles/App.css
@@ -1,0 +1,29 @@
+/* :root {
+  --color-blue-light: rgb(197, 197, 255);
+} */
+
+* {
+  margin: 0;
+  padding: 0;
+}
+
+input,
+select {
+  outline: none;
+}
+
+.app {
+  block-size: 100%;
+  min-block-size: 100vh;
+  background-color: rgb(197, 197, 255);
+}
+
+.app__splitted {
+  display: flex;
+}
+
+.error-message {
+  text-align: center;
+  color: red;
+  margin-block-start: 30px;
+}
