+++ src/types/AnimeData.ts
